options window on main context
+options window now is rendered to main context instead of a new one
+audio player now displays MPEG album cover images on the main player surface
+playlist info rendering bug fixes
+option window bug fixes

+++ demp/src/audio/mp_channel_attrib.h
@@ -49,7 +49,7 @@ namespace Audio
 			std::string format;
 
 			f32 bitrate;
-			f32 channels;
+			s32 channels;
 			f32 freq;
 			f32 size;	// in bytes
 			f32 length;	// in sec
